+++ source/Public/New-DscResourcePowerShellHelp.ps1
@@ -1,47 +1,81 @@
 <#
     .SYNOPSIS
         New-DscResourcePowerShellHelp generates PowerShell compatible help files
-        for a DSC resource module
+        for a DSC resource module.
 
     .DESCRIPTION
-        The New-DscResourcePowerShellHelp cmdlet will review all of the MOF based
-        resources in a specified module directory and will inject PowerShell help
-        files for each resource. These help files include details on the property
-        types for each resource, as well as a text description and examples where
-        they exist.
+        The New-DscResourcePowerShellHelp generates PowerShell compatible help files
+        for a DSC resource module.
+
+        The command will review all of the MOF-based and class-based resources in a
+        specified module directory and will inject PowerShell help files for each
+        resource. These help files include details on the property types for each
+        resource,  as well as a text description and examples where they exist.
 
         The help files are output to the OutputPath directory if specified. If not,
         they are output to the relevant resource's 'en-US' directory either in the
         path set by 'ModulePath' or to 'DestinationModulePath' if set.
 
-        A README.md with a text description must exist in the resource's subdirectory
-        for the help file to be generated. To get examples added to the conceptual
-        help the examples must be present in an individual resource example folder,
-        e.g. 'Examples/Resources/<ResourceName>/1-Example.ps1'. Prefixing the value
+        For MOF-based resources a README.md with a text description must exist in
+        the resource's subdirectory for the help file to be generated. For class-based
+        resources each DscResource should have their own file in the Classes folder
+        (using the template of the Sampler project).
+
+        To get examples added to the conceptual help the examples must be present
+        in an individual resource example folder, e.g.
+        'Examples/Resources/<ResourceName>/1-Example.ps1'. Prefixing the value
         with a number will sort the examples in that order.
+
+        Example directory structure:
+
+            Examples
+                \---Resources
+                    \---MyResourceName
+                            1-FirstExample.ps1
+                            2-SecondExample.ps1
+                            3-ThirdExample.ps1
 
         These help files can then be read by passing the name of the resource as a
         parameter to Get-Help.
 
     .PARAMETER ModulePath
-        The path to the root of the DSC resource module (where the PSD1 file is
-        found). In this path the folder DSCResources must be present.
+        The path to the root of the DSC resource module where the PSD1 file is
+        found, for example the folder 'source'. If there are MOF-based resources
+        there should be a 'DSCResources' child folder in this path. If using
+        class-based resources there should be a 'Classes' child folder in this path.
 
     .PARAMETER DestinationModulePath
         The destination module path can be used to set the path where module is
         built before being deployed. This must be set to the root of the built
         module, e.g 'c:\repos\ModuleName\output\ModuleName\1.0.0'.
-        If OutputPath is also used at the same time that will override this path.
+
+        The conceptual help file will be saved in this path. For MOF-based resources
+        it will be saved to the 'en-US' folder that is inside in either the 'DSCResources'
+        or 'DSCClassResources' folder (if using that pattern for class-based resources).
+
+        When using the pattern with having all powershell classes in the same
+        module script file (.psm1) and all class-based resource are found in that
+        file (not using 'DSCClassResources'). This path will be used to find the
+        built module when generating conceptual help for class-based resource.
+        It will also be used to save the conceptual help to the built modules
+        'en-US' folder.
+
+        If OutputPath is assigned that will be used for saving the output instead.
 
     .PARAMETER OutputPath
         The output path can be used to set the path where all the generated files
         will be saved (all files to the same path).
 
     .PARAMETER MarkdownCodeRegularExpression
-        An array of regular expressions that should be used to parse the parameter
-        descriptions in the schema MOF. Each regular expression must be written
-        so that the capture group 0 is the full match and the capture group 1 is
-        the text that should be kept.
+        An array of regular expressions that should be used to parse the text of
+        the synopsis, description and parameter descriptions. Each regular expression
+        must be written so that the capture group 0 is the full match and the capture
+        group 1 is the text that should be kept. This is meant to be used to remove
+        markdown code, but it can be used for anything as it follow the previous
+        mention pattern for the regular expression sequence.
+
+    .PARAMETER Force
+        When set the to $true and existing conceptual help file will be overwritten.
 
     .EXAMPLE
         New-DscResourcePowerShellHelp -ModulePath C:\repos\SharePointDsc
@@ -82,11 +116,16 @@ function New-DscResourcePowerShellHelp
 
         [Parameter()]
         [System.String[]]
-        $MarkdownCodeRegularExpression = @()
+        $MarkdownCodeRegularExpression = @(),
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Force
     )
 
+    #region MOF-based resource
     $mofSearchPath = (Join-Path -Path $ModulePath -ChildPath '\**\*.schema.mof')
-    $mofSchemas = @(Get-ChildItem -Path $mofSearchPath -Recurse)
+    $mofSchemas = @(Get-ChildItem -Path $mofSearchPath -File -Recurse)
 
     Write-Verbose -Message ($script:localizedData.FoundMofFilesMessage -f $mofSchemas.Count, $ModulePath)
 
@@ -121,6 +160,11 @@ function New-DscResourcePowerShellHelp
             $descriptionContent = $descriptionContent -replace '\r\n\s{4}\r\n', "`r`n`r`n"
             $descriptionContent = $descriptionContent -replace '\s{4}$', ''
 
+            if (-not [System.String]::IsNullOrEmpty($descriptionContent))
+            {
+                $descriptionContent = Get-RegularExpressionParsedText -Text $descriptionContent -RegularExpression $MarkdownCodeRegularExpression
+            }
+
             $output += $descriptionContent
             $output += "`r`n"
 
@@ -143,58 +187,23 @@ function New-DscResourcePowerShellHelp
                     $output += "`r`n"
                 }
 
-                $propertyDescription = $property.Description
-
-                if ($MarkdownCodeRegularExpression.Count -gt 0)
+                if (-not [System.String]::IsNullOrEmpty($property.Description))
                 {
-                    foreach ($parseRegularExpression in $MarkdownCodeRegularExpression)
-                    {
-                        $allMatches = $propertyDescription | Select-String -Pattern $parseRegularExpression -AllMatches
-
-                        foreach ($regularExpressionMatch in $allMatches.Matches)
-                        {
-                            <#
-                                Always assume the group 0 is the full match and
-                                the group 1 contain what we should replace with.
-                            #>
-                            $propertyDescription = $propertyDescription -replace @(
-                                [RegEx]::Escape($regularExpressionMatch.Groups[0].Value),
-                                $regularExpressionMatch.Groups[1].Value
-                            )
-                        }
-                    }
+                    $propertyDescription = Get-RegularExpressionParsedText -Text $property.Description -RegularExpression $MarkdownCodeRegularExpression
                 }
 
-                $output += "    " + $propertyDescription
+                $output += "    {0}" -f $propertyDescription
                 $output += "`r`n`r`n"
             }
 
-            $exampleSearchPath = "\Examples\Resources\$($result.FriendlyName)\*.ps1"
-            $examplesPath = (Join-Path -Path $ModulePath -ChildPath $exampleSearchPath)
-            $exampleFiles = @(Get-ChildItem -Path $examplesPath -ErrorAction SilentlyContinue)
+            $resourceExamplePath = Join-Path -Path $ModulePath -ChildPath ('Examples\Resources\{0}' -f $result.FriendlyName)
 
-            if ($exampleFiles.Count -gt 0)
-            {
-                $exampleCount = 1
+            $exampleContent = Get-ResourceExampleAsText -Path $resourceExamplePath
 
-                Write-Verbose -Message "Found $($exampleFiles.count) Examples for resource $($result.FriendlyName)"
+            $output += $exampleContent
 
-                foreach ($exampleFile in $exampleFiles)
-                {
-                    $exampleContent = Get-DscResourceHelpExampleContent `
-                        -ExamplePath $exampleFile.FullName `
-                        -ExampleNumber ($exampleCount++)
-
-                    $exampleContent = $exampleContent -replace '\r?\n', "`r`n"
-
-                    $output += $exampleContent
-                    $output += "`r`n"
-                }
-            }
-            else
-            {
-                Write-Warning -Message ($script:localizedData.NoExampleFileFoundWarning -f $result.FriendlyName)
-            }
+            # Trim excessive blank lines and indents at the end.
+            $output = $output -replace '[\r|\n|\s]+$', "`r`n"
 
             $outputFileName = "about_$($result.FriendlyName).help.txt"
 
@@ -212,8 +221,8 @@ function New-DscResourcePowerShellHelp
 
                 $savePath = Join-Path -Path $DestinationModulePath -ChildPath $dscRootFolderName |
                     Join-Path -ChildPath $resourceRelativePath |
-                    Join-Path -ChildPath 'en-US' |
-                    Join-Path -ChildPath $outputFileName
+                        Join-Path -ChildPath 'en-US' |
+                            Join-Path -ChildPath $outputFileName
             }
             else
             {
@@ -224,11 +233,210 @@ function New-DscResourcePowerShellHelp
 
             Write-Verbose -Message ($script:localizedData.OutputHelpDocumentMessage -f $savePath)
 
-            $output | Out-File -FilePath $savePath -Encoding 'ascii' -Force
+            $output | Out-File -FilePath $savePath -Encoding 'ascii' -Force:$Force
         }
         else
         {
             Write-Warning -Message ($script:localizedData.NoDescriptionFileFoundWarning -f $result.FriendlyName)
         }
     }
+    #endregion MOF-based resource
+
+    #region Class-based resource
+    if (-not [System.String]::IsNullOrEmpty($DestinationModulePath) -and (Test-Path -Path $DestinationModulePath))
+    {
+        $getChildItemParameters = @{
+            Path        = Join-Path -Path $DestinationModulePath -ChildPath '*'
+            Include     = '*.psm1'
+            ErrorAction = 'Stop'
+            File        = $true
+            Recurse     = $true
+        }
+
+        $builtModuleScriptFiles = Get-ChildItem @getChildItemParameters
+
+        # Looping through each module file (normally just one).
+        foreach ($builtModuleScriptFile in $builtModuleScriptFiles)
+        {
+            $tokens, $parseErrors = $null
+
+            $ast = [System.Management.Automation.Language.Parser]::ParseFile($builtModuleScriptFile.FullName, [ref] $tokens, [ref] $parseErrors)
+
+            if ($parseErrors)
+            {
+                throw $parseErrors
+            }
+
+            $astFilter = {
+                $args[0] -is [System.Management.Automation.Language.TypeDefinitionAst] `
+                    -and $args[0].IsClass -eq $true `
+                    -and $args[0].Attributes.Extent.Text -imatch '\[DscResource\(.*\)\]'
+            }
+
+            $dscResourceAsts = $ast.FindAll($astFilter, $true)
+
+            Write-Verbose -Message ($script:localizedData.FoundClassBasedMessage -f $dscResourceAsts.Count, $builtModuleScriptFile.FullName)
+
+            # Looping through each class-based resource.
+            foreach ($dscResourceAst in $dscResourceAsts)
+            {
+                Write-Verbose -Message ($script:localizedData.GenerateHelpDocumentMessage -f $dscResourceAst.Name)
+
+                <#
+                    PowerShell classes does not support comment-based help. There is
+                    no GetHelpContent() on the TypeDefinitionAst.
+
+                    We use the ScriptBlockAst to filter out our class-based resource
+                    script block from the source file and use that to get the
+                    comment-based help.
+                #>
+                $sourceFilePath = Join-Path -Path $ModulePath -ChildPath ('Classes/*{0}.ps1' -f $dscResourceAst.Name)
+                $sourceFilePath = Resolve-Path -Path $sourceFilePath
+
+                Write-Verbose -Message ($script:localizedData.ClassBasedCommentBasedHelpMessage -f $sourceFilePath)
+
+                $ast = [System.Management.Automation.Language.Parser]::ParseFile($sourceFilePath, [ref] $tokens, [ref] $parseErrors)
+
+                $dscResourceCommentBasedHelp = $ast.GetHelpContent()
+
+                $synopsis = $dscResourceCommentBasedHelp.Synopsis
+                $synopsis = $synopsis -replace '[\r|\n]+$' # Removes all blank rows at the end
+                $synopsis = $synopsis -replace '\r?\n', "`r`n" # Normalize to CRLF
+                $synopsis = $synopsis -replace '\r\n', "`r`n    " # Indent all rows
+                $synopsis = $synopsis -replace '[ ]+\r\n', "`r`n" # Remove indentation from blank rows
+
+                $description = $dscResourceCommentBasedHelp.Description
+                $description = $description -replace '[\r|\n]+$' # Removes all blank rows and whitespace at the end
+                $description = $description -replace '\r?\n', "`r`n" # Normalize to CRLF
+                $description = $description -replace '\r\n', "`r`n    " # Indent all rows
+                $description = $description -replace '[ ]+\r\n', "`r`n" # Remove indentation from blank rows
+
+                $output = ".NAME`r`n"
+                $output += '    {0}' -f $dscResourceAst.Name
+                $output += "`r`n`r`n"
+                $output += ".SYNOPSIS`r`n"
+
+                if (-not [System.String]::IsNullOrEmpty($synopsis))
+                {
+                    $synopsis = Get-RegularExpressionParsedText -Text $synopsis -RegularExpression $MarkdownCodeRegularExpression
+
+                    $output += '    {0}' -f $synopsis
+                }
+
+                $output += "`r`n`r`n"
+                $output += ".DESCRIPTION`r`n"
+
+                if (-not [System.String]::IsNullOrEmpty($description))
+                {
+                    $description = Get-RegularExpressionParsedText -Text $description -RegularExpression $MarkdownCodeRegularExpression
+
+                    $output += '    {0}' -f $description
+                }
+
+                $output += "`r`n`r`n"
+
+                $astFilter = {
+                    $args[0] -is [System.Management.Automation.Language.PropertyMemberAst]
+                }
+
+                $propertyMemberAsts = $dscResourceAst.FindAll($astFilter, $true)
+
+                # Looping through each resource property.
+                foreach ($propertyMemberAst in $propertyMemberAsts)
+                {
+                    Write-Verbose -Message ($script:localizedData.FoundClassResourcePropertyMessage -f $propertyMemberAst.Name, $dscResourceAst.Name)
+
+                    $astFilter = {
+                        $args[0] -is [System.Management.Automation.Language.NamedAttributeArgumentAst]
+                    }
+
+                    $propertyNamedAttributeArgumentAsts = $propertyMemberAst.FindAll($astFilter, $true)
+
+                    $isKeyProperty = 'Key' -in $propertyNamedAttributeArgumentAsts.ArgumentName
+                    $isMandatoryProperty = 'Mandatory' -in $propertyNamedAttributeArgumentAsts.ArgumentName
+                    $isReadProperty = 'NotConfigurable' -in $propertyNamedAttributeArgumentAsts.ArgumentName
+
+                    if ($isKeyProperty)
+                    {
+                        $propertyState = 'Key'
+                    }
+                    elseif ($isMandatoryProperty)
+                    {
+                        $propertyState = 'Required'
+                    }
+                    elseif ($isReadProperty)
+                    {
+                        $propertyState = 'Read'
+                    }
+                    else
+                    {
+                        $propertyState = 'Write'
+                    }
+
+                    $output += ".PARAMETER {0}`r`n" -f $propertyMemberAst.Name
+                    $output += '    {0} - {1}' -f $propertyState, $propertyMemberAst.PropertyType.TypeName.FullName
+                    $output += "`r`n"
+
+                    $astFilter = {
+                        $args[0] -is [System.Management.Automation.Language.AttributeAst] `
+                            -and $args[0].TypeName.Name -eq 'ValidateSet'
+                    }
+
+                    $propertyAttributeAsts = $propertyMemberAst.FindAll($astFilter, $true)
+
+                    if ($propertyAttributeAsts)
+                    {
+                        $output += "    Allowed values: {0}" -f ($propertyAttributeAsts.PositionalArguments.Value -join ', ')
+                        $output += "`r`n"
+                    }
+
+                    # The key name must be upper-case for it to match the right item in the list of parameters.
+                    $propertyDescription = ($dscResourceCommentBasedHelp.Parameters[$propertyMemberAst.Name.ToUpper()] -replace '[\r|\n]+$')
+
+                    $propertyDescription = $propertyDescription -replace '[\r|\n]+$' # Removes all blank rows at the end
+                    $propertyDescription = $propertyDescription -replace '\r?\n', "`r`n" # Normalize to CRLF
+                    $propertyDescription = $propertyDescription -replace '\r\n', "`r`n    " # Indent all rows
+                    $propertyDescription = $propertyDescription -replace '[ ]+\r\n', "`r`n" # Remove indentation from blank rows
+
+                    if (-not [System.String]::IsNullOrEmpty($propertyDescription))
+                    {
+                        $propertyDescription = Get-RegularExpressionParsedText -Text $propertyDescription -RegularExpression $MarkdownCodeRegularExpression
+
+                        $output += "    {0}" -f $propertyDescription
+                        $output += "`r`n"
+                    }
+
+                    $output += "`r`n"
+                }
+
+                $examplesPath = Join-Path -Path $ModulePath -ChildPath ('Examples\Resources\{0}' -f $dscResourceAst.Name)
+
+                $exampleContent = Get-ResourceExampleAsText -Path $examplesPath
+
+                $output += $exampleContent
+
+                # Trim excessive blank lines and indents at the end, then insert a last blank line.
+                $output = $output -replace '[\r?\n|\s]+$', "`r`n"
+
+                $outputFileName = 'about_{0}.help.txt' -f $dscResourceAst.Name
+
+                if ($PSBoundParameters.ContainsKey('OutputPath'))
+                {
+                    # Output to $OutputPath if specified.
+                    $savePath = Join-Path -Path $OutputPath -ChildPath $outputFileName
+                }
+                else
+                {
+                    # Output to the built modules en-US folder.
+                    $savePath = Join-Path -Path $DestinationModulePath -ChildPath 'en-US' |
+                        Join-Path -ChildPath $outputFileName
+                }
+
+                Write-Verbose -Message ($script:localizedData.OutputHelpDocumentMessage -f $savePath)
+
+                $output | Out-File -FilePath $savePath -Encoding 'ascii' -NoNewLine -Force:$Force
+            }
+        }
+    }
+    #endregion Class-based resource
 }